Challenge 4: This is good co-primen.
Factorizing the input number into primes, then using a sieve
weeding out the numbers which aren't co-prime leads to a
program which is faster than this solution, which just calculates
the gcd for all numbers less than the target number.

But calculating the gcd is much simpler to program, and requires
less memory. And it's fast enough for a one-shot program.