Challenge 8.
This does not calculate the correct solution.

In fact, I'm not even to reproduce the example; I've no idea
how milk can both expire and be drunk. The description says
that "Milk always expires on the 5th day after you buy it,
before you use it that morning." Which would suggest that
on Day 6, the milk we bought on Day 1 has expired, so we
drink the milk we bought on Day 5, leaving us with 900 ml.